Refactoring: Add submodules for geom
