+++ src/geom/mod.rs
@@ -1,0 +1,12 @@
+mod camera;
+mod rect;
+mod screen;
+mod transforms;
+
+pub use camera::Camera;
+pub use rect::Rect;
+pub use screen::Screen;
+pub use transforms::{
+    matrix3_to_array, scale_rotate_translate, scale_translate, scale_translate3,
+    translate_rotate_scale,
+};
